remove old generator functions and use number representation
