+++ elasticai/creator/tests/vhdl/test_generate_lstm_cell_testbench_vhd.py
@@ -188,8 +188,74 @@ end architecture rtl;
             input_size=5,
             hidden_size=20,
             component_name="lstm_cell",
-            test_x_h_data='x"018a",x"ffb5",x"fdd3",x"0091",x"feeb",x"0099",x"fe72",x"ffa9",x"01da",x"ffc9",x"ff42",x"0090",x"0042",x"ffd4",x"ff53",x"00f0",x"007d",x"0134",x"0015",x"fecd",x"ffff",x"ff7c",x"ffb2",x"fe6c",x"01b4",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000"',
-            test_c_data='x"0034",x"ff8d",x"ff6e",x"ff72",x"fee0",x"ffaf",x"fee9",x"ffeb",x"ffe9",x"00af",x"ff2a",x"0000",x"ff40",x"002f",x"009f",x"00a3",x"ffc2",x"024d",x"fe1f",x"fff4",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000",x"0000"',
+            test_x_h_data=[
+                "018a",
+                "ffb5",
+                "fdd3",
+                "0091",
+                "feeb",
+                "0099",
+                "fe72",
+                "ffa9",
+                "01da",
+                "ffc9",
+                "ff42",
+                "0090",
+                "0042",
+                "ffd4",
+                "ff53",
+                "00f0",
+                "007d",
+                "0134",
+                "0015",
+                "fecd",
+                "ffff",
+                "ff7c",
+                "ffb2",
+                "fe6c",
+                "01b4",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+            ],
+            test_c_data=[
+                "0034",
+                "ff8d",
+                "ff6e",
+                "ff72",
+                "fee0",
+                "ffaf",
+                "fee9",
+                "ffeb",
+                "ffe9",
+                "00af",
+                "ff2a",
+                "0000",
+                "ff40",
+                "002f",
+                "009f",
+                "00a3",
+                "ffc2",
+                "024d",
+                "fe1f",
+                "fff4",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+                "0000",
+            ],
             h_out=[
                 34,
                 -80,
